[PiranhaJava] Add configuration for test method cleanup
